requirements.txt
# Core LlamaIndex dependencies
llama-index==0.9.48
llama-index-llms-openai==0.1.6
llama-index-embeddings-openai==0.1.6
llama-index-readers-file==0.1.6
llama-index-vector-stores-chroma==0.1.6

# Web framework and API
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-login==0.6.3

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Security and validation
bcrypt==4.1.2
python-jose==3.3.0
marshmallow==3.20.1

# Database and storage
chromadb==0.4.22
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Optional: For better performance
numpy==1.24.3
pandas==2.0.3 